Fix cross_chain.rs panic on short right_id input (line 95) Fix Bitcoin source transfer Merkle/header placeholder issue (cross_chain_impl.rs line 71/247) Fix bypassable finality check in CLI flow (cross_chain.rs line 167, cross_chain_impl.rs line 283) Fix core validation accepting proof/chain-type mismatch (client.rs line 354) Replace synthetic commitments with anchored data in client.rs (line 235) Remove mocked MCP tool stubs (proof.ts line 23, transfer.ts line 91) Fix Ethereum balance overflow (u64 for wei) in csv-wallet chain_api.rs Fix local-dev typo (sourceTXHash) in csv-local-dev registry.js Fix hardcoded Mainnet and total_indexed_blocks: 0 in explorer sync.rs Add simulation vs production pathway markers in CLI/MCP outputs Fix as_any unimplemented! panic in ethereum rpc.rs

+++ csv-wallet/src/services/chain_api.rs
@@ -243,8 +243,9 @@ impl ChainApi {
             .and_then(|v| v.as_str())
             .ok_or_else(|| ChainApiError::ApiError("Missing result in response".to_string()))?;
 
-        // Parse hex balance (in wei) to f64 (in ETH)
-        let balance_wei = u64::from_str_radix(balance_hex.trim_start_matches("0x"), 16)
+        // Parse hex balance (in wei) to f64 (in ETH).
+        // Use u128 to avoid overflow for balances > 18.4 ETH.
+        let balance_wei = u128::from_str_radix(balance_hex.trim_start_matches("0x"), 16)
             .map_err(|e| ChainApiError::ApiError(format!("Invalid hex balance: {}", e)))?;
 
         Ok(balance_wei as f64 / 1e18)
